Refactoring of timer event names
