+++ activiti-api-process-model-impl/src/main/java/org/activiti/api/runtime/event/impl/BPMNTimerExecutedEventImpl.java
@@ -1,0 +1,25 @@
+package org.activiti.api.runtime.event.impl;
+
+import org.activiti.api.process.model.BPMNTimer;
+import org.activiti.api.process.model.events.BPMNTimerEvent;
+import org.activiti.api.process.model.events.BPMNTimerExecutedEvent;
+
+public class BPMNTimerExecutedEventImpl extends RuntimeEventImpl<BPMNTimer, BPMNTimerEvent.TimerEvents> implements BPMNTimerExecutedEvent {
+
+    public BPMNTimerExecutedEventImpl() {
+    }
+
+    public BPMNTimerExecutedEventImpl(BPMNTimer entity) {
+        super(entity);
+    }
+
+    @Override
+    public TimerEvents getEventType() {
+        return BPMNTimerEvent.TimerEvents.JOB_EXECUTION_SUCCESS;
+    }
+
+    @Override
+    public String toString() {
+        return "BPMNTimerExecutedEventImpl{" + super.toString() + "}";
+    }
+}
