+++ activiti-api-process-model/src/main/java/org/activiti/api/process/model/events/BPMNTimerCancelledEvent.java
@@ -16,6 +16,6 @@
 
 package org.activiti.api.process.model.events;
 
-public interface BPMNTimerCanceledEvent extends BPMNTimerEvent {
+public interface BPMNTimerCancelledEvent extends BPMNTimerEvent {
 
 }
